requirements.txt
flask~=3.1.1
WTForms~=3.2.1
SQLAlchemy~=2.0.41
email_validator>=1.0.0

Flask-Login~=0.6.3
Werkzeug~=3.1.3
python-dotenv~=1.1.1
jdatetime~=5.2.0